34 remove submissions

+++ src/regtech_cleanup_api/routers/filing_cleanup.py
@@ -21,9 +21,7 @@ from regtech_cleanup_api.services.validation import is_valid_cleanup_lei
 logger = logging.getLogger(__name__)
 
 
-def set_db(
-    request: Request, session: Annotated[Session, Depends(get_filing_session)]
-):
+def set_db(request: Request, session: Annotated[Session, Depends(get_filing_session)]):
     request.state.db_session = session
 
 
@@ -72,6 +70,51 @@ def delete_filing(request: Request, lei: str, period_code: str):
                     detail="Failed to delete filing data",
                 ) from e
 
+            try:
+                repo.delete_user_actions(session, user_action_ids)
+            except Exception as e:
+                raise RegTechHttpException(
+                    status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
+                    name="User Action Delete Failed",
+                    detail="Failed to delete user action data",
+                ) from e
+
+            delete_from_storage(period_code, lei)
+
+        except Exception as e:
+            raise e
+    else:
+        raise RegTechHttpException(
+            status_code=HTTPStatus.NOT_ACCEPTABLE,
+            name="Invalid LEI",
+            detail="Not a valid LEI",
+        )
+    return Response(status_code=status.HTTP_202_ACCEPTED)
+
+
+@router.delete("/institutions/{lei}/submissions/{period_code}")
+def delete_submissions(request: Request, lei: str, period_code: str):
+    if is_valid_cleanup_lei(lei):
+        try:
+            session = request.state.db_session
+            try:
+                user_action_ids = repo.get_user_action_ids(
+                    session, lei=lei, period_code=period_code, submissions=True
+                )
+            except Exception as e:
+                raise RegTechHttpException(
+                    status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
+                    name="Missing User Action Data",
+                    detail="Failed to get user action data",
+                ) from e
+            try:
+                repo.delete_submissions(session, lei, period_code)
+            except Exception as e:
+                raise RegTechHttpException(
+                    status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
+                    name="Submission Delete Failed",
+                    detail="Failed to delete submission data",
+                ) from e
             try:
                 repo.delete_user_actions(session, user_action_ids)
             except Exception as e:
